+++ src/WhereGroup/MetadorBundle/Controller/DefaultController.php
@@ -14,11 +14,12 @@ use WhereGroup\MetadorBundle\Entity\Metadata;
 use WhereGroup\MetadorBundle\Entity\Helptext;
 use WhereGroup\MetadorBundle\Entity\Address;
 use WhereGroup\MetadorBundle\Event\MetadataChangeEvent;
+use WhereGroup\MetadorBundle\Component\MetadorController;
 
 /**
  * @Route("/metador")
  */
-class DefaultController extends Controller
+class DefaultController extends MetadorController
 {
     /**
      * @Route("/")
@@ -27,47 +28,11 @@ class DefaultController extends Controller
     public function indexAction() {
         $limit = 30;
         $offset = 0;
-        $em = $this->getDoctrine()->getManager();
-        $qb = $em->createQueryBuilder();
-
-        $service = $em->createQueryBuilder('m')->select('m.id, m.title')
-                ->from('WhereGroupMetadorBundle:Metadata','m')
-                ->where($qb->expr()->orx(
-                    $qb->expr()->eq('m.hierarchyLevel', '?1')
-                ))
-                ->orderBy('m.updateTime', 'DESC')
-                ->setFirstResult( $offset )
-                ->setMaxResults( $limit )
-                ->setParameters(array(1 => 'service'))
-                ->getQuery()
-                ->getResult();
-
-
-        $dataset = $em->createQueryBuilder('x')->select('x.id, x.title')
-                ->from('WhereGroupMetadorBundle:Metadata','x')
-                ->where($qb->expr()->orx(
-                    $qb->expr()->eq('x.hierarchyLevel', '?1'),
-                    $qb->expr()->eq('x.hierarchyLevel', '?2')
-                ))
-                ->orderBy('x.updateTime', 'DESC')
-                ->setFirstResult( $offset )
-                ->setMaxResults( $limit )
-                ->setParameters(array(
-                    1 => 'dataset',
-                    2 => 'series',
-                ))
-                ->getQuery()
-                ->getResult();
-
-        $address = $em->createQueryBuilder('y')->select('y.id, y.individualName')
-                ->from('WhereGroupMetadorBundle:Address','y')
-                ->getQuery()
-                ->getResult();
 
         return array(
-            'service' => $service,
-            'dataset' => $dataset,
-            'address' => $address
+            'service' => $this->getService($limit, $offset),
+            'dataset' => $this->getDataset($limit, $offset),
+            'address' => $this->getAddress()
         );
     }
 
@@ -80,7 +45,7 @@ class DefaultController extends Controller
         $metadata = $em->getRepository('WhereGroupMetadorBundle:Metadata')->findOneById($id);
         
         if($metadata) {
-            $p = unserialize($metadata->getMetadata());
+            $p = $metadata->getObject();
 
             $data = array('p' => $p);
 
@@ -123,7 +88,7 @@ class DefaultController extends Controller
         $metadata = $em->getRepository('WhereGroupMetadorBundle:Metadata')->findOneById($id);
         
         if($metadata) {
-            $p = unserialize($metadata->getMetadata());
+            $p = $metadata->getObject();
 
             die('<pre>' . print_r($p, 1) . '</pre>');