+++ app/Resources/doc/changelog.md
@@ -3,6 +3,7 @@
 ## 2.0.4.0
 * Helptext import/export commands.
 * Added Searchbundle.
+* Changed metadata permissions.
 
 ## 2.0.3.3
 * Minor bugfixes.